Add WebAssembly support for Lean4Lean typechecker and related tests
- Implement lean4lean_typechecker_shims.c for Lean4Lean integration with WASM.
- Create run.mjs to execute the WASM module and validate Lean4Wasm output.
- Add run_test.sh for smoke testing the Lean4Lean WASM compilation and execution.
- Introduce smoke_lean4lean_typechecker.mjs to test Lean4Lean typechecking functionality.
- Ensure proper error handling and validation of Lean4Lean ASTs in the new tests.

+++ tests/freestanding_runtime/test_cow.c
@@ -19,7 +19,7 @@
 /* --- Freestanding runtime public API surface we exercise. ---------- */
 
 extern void *lean_box(size_t n);
-extern uint64_t lean_unbox(void *o);
+extern size_t lean_unbox(void *o);
 
 extern void  lean_inc_ref(void *o);
 extern void  lean_dec_ref(void *o);
